+++ src/main/kotlin/us/frollo/frollosdk/model/api/contacts/ContactResponse.kt
@@ -1,6 +1,23 @@
+/*
+ * Copyright 2019 Frollo
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
 package us.frollo.frollosdk.model.api.contacts
 
 import com.google.gson.annotations.SerializedName
+import us.frollo.frollosdk.model.coredata.aggregation.providers.AggregatorType
 import us.frollo.frollosdk.model.coredata.contacts.PaymentDetails
 import us.frollo.frollosdk.model.coredata.contacts.PaymentMethod
 
@@ -13,6 +30,9 @@ internal data class ContactResponse(
     @SerializedName("name") val name: String,
     @SerializedName("nick_name") val nickName: String,
     @SerializedName("description") val description: String?,
+    @SerializedName("aggregator") val aggregatorType: AggregatorType,
+    @SerializedName("consent_id") val consentId: Long?,
+    @SerializedName("editable") val editable: Boolean,
     @SerializedName("payment_method") val paymentMethod: PaymentMethod,
 
     // DO NOT add @SerializedName("payment_details") to this field as it cannot be directly