+++ src/main/kotlin/us/frollo/frollosdk/model/coredata/contacts/DigitalWalletType.kt
@@ -16,31 +16,22 @@
 
 package us.frollo.frollosdk.model.coredata.contacts
 
+import androidx.annotation.Keep
 import com.google.gson.annotations.SerializedName
 import us.frollo.frollosdk.extensions.serializedName
 
-/**
- * Indicates the payment method type for the contact
- */
-enum class PaymentMethod {
+/** Type of Digital Wallet Identifier */
+@Keep
+enum class DigitalWalletType {
 
-    /** Pay Anyone type*/
-    @SerializedName("pay_anyone") PAY_ANYONE,
+    /** Contact Name */
+    @SerializedName("CONTACT_NAME") CONTACT_NAME,
 
-    /** Pay ID type */
-    @SerializedName("pay_id") PAY_ID,
+    /** Email */
+    @SerializedName("EMAIL") EMAIL,
 
-    /** BPay type */
-    @SerializedName("bpay") BPAY,
-
-    /** International Payment type */
-    @SerializedName("international") INTERNATIONAL,
-
-    /** Digital Wallet Payment type */
-    @SerializedName("digital_wallet") DIGITAL_WALLET,
-
-    /** Card Payment type */
-    @SerializedName("card") CARD;
+    /** Phone number */
+    @SerializedName("TELEPHONE") TELEPHONE;
 
     /** Enum to serialized string */
     // This override MUST be used for this enum to work with Retrofit @Path or @Query parameters