+++ src/main/kotlin/us/frollo/frollosdk/network/deserializer/ContactResponseDeserializer.kt
@@ -26,6 +26,8 @@ internal object ContactResponseDeserializer : JsonDeserializer<ContactResponse> 
                         PaymentMethod.BPAY -> Gson().fromJson(elem, PaymentDetails.Biller::class.java)
                         PaymentMethod.PAY_ID -> Gson().fromJson(elem, PaymentDetails.PayID::class.java)
                         PaymentMethod.INTERNATIONAL -> Gson().fromJson(elem, PaymentDetails.International::class.java)
+                        PaymentMethod.DIGITAL_WALLET -> Gson().fromJson(elem, PaymentDetails.DigitalWallet::class.java)
+                        PaymentMethod.CARD -> Gson().fromJson(elem, PaymentDetails.Card::class.java)
                     }
                 }
             }